requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Supabase client
supabase==2.7.4

# HTTP client (for calling n8n webhooks)
httpx==0.27.2

# Config management
pydantic-settings==2.4.0

# Phase 3 dependencies (install now to avoid redeployment)
# Embeddings + document parsing
openai==1.51.0
PyMuPDF==1.24.10
python-docx==1.1.2
openpyxl==3.1.5
langchain-text-splitters==0.3.0